chore: do strict unmarshaling of configuration
Unmarshaling of types that implement custom unmarshaling interface will not use that option,
which in case of this configuration will lead to basically loosing that setting.

https://github.com/go-yaml/yaml/issues/460

This change adds strict unmarshaling, though in not ideal way.

Also add test that tries to load example configurations as first test for their validity.