feat: setup watchdog (#93)
* feat: setup watchdog

It can be a temporary solution to situations where
unknown issue will lock up SoC resulting
in values not being updated and not responding to commands.

"Production" configuration will already automatically reset SoC on assert fail,
but issue may not always be a failing assert,
but a memory related issue, for example/

* add configuration for watchdog

Enabled by default.

It is always disabled for debug configuration,
but can also be disabled manually
via config option.