feat!: prefer Zigbee add on SDK (#94)
* chore: move some debug configuration to Kconfig

This will limit options that need to be set manually

* feat!: prefer Zigbee add on SDK

Breaking: better naming and placement for NCS configuration options.
Should provide less confusion for people that need to use this options.

This change will prefer the SDKs that contain Zigbee Add On.
As nRF Connect removed Zigbee from SDK v3.0.0 and forward
we need to use the new Zigbee Add On SDKs.