feat: Support nRF54L15 (DK) (#101)
* feat: add run context

This will allow running commands with the same context.
Could be useful to run multiple commands,
for example when building & flashing.

* chore: update workdir resolution

* chore: detect SoC on config parse

This will use board qualifier, or "nrf52840" by default.

While it can also be done via file parsing, the SoC is generally defined in the board mame, especially in later versions of Zephyr, so no extra work is needed.

* feat: select proper timer for nrf54l series

* feat: support nRF54L all three GPIO ports

* fix: remove unknown config option

* fix: on_off template cpp syntax